#458 Separate `SqlService` and `JpaRepository` contracts
