+++ hartshorn-persistence/src/main/java/org/dockbox/hartshorn/persistence/SqlService.java
@@ -17,30 +17,15 @@
 
 package org.dockbox.hartshorn.persistence;
 
-import org.dockbox.hartshorn.core.domain.Exceptional;
 import org.dockbox.hartshorn.core.context.ContextCarrier;
 import org.dockbox.hartshorn.core.properties.AttributeHolder;
-
-import java.util.Set;
 
 import javax.persistence.EntityManager;
 
 /**
  * Low-level interface for communicating with SQL databases.
  */
-public interface SqlService extends AttributeHolder, ContextCarrier, EntityManager {
-
-    void save(Object object);
-
-    void update(Object object);
-
-    void updateOrSave(Object object);
-
-    void delete(Object object);
-
-    <T> Set<T> findAll(Class<T> type);
-
-    <T> Exceptional<T> findById(Class<T> type, Object id);
+public interface SqlService extends AttributeHolder, ContextCarrier, EntityManager, JpaRepository {
 
     EntityManager entityManager();
 }
